requirements.txt
streamlit>=1.30.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pandas>=2.0.0
pytest>=8.0.0
pydantic>=2.0.0
